refactor: moved appendLeaves functionality to TreeBase
