+++ yarn-project/merkle-tree/src/tree_base.ts
@@ -6,7 +6,7 @@ import { toBigIntLE, toBufferLE } from '@aztec/foundation/bigint-buffer';
 
 const MAX_DEPTH = 254;
 
-export const indexToKeyHash = (name: string, level: number, index: bigint) => `${name}:${level}:${index}`;
+const indexToKeyHash = (name: string, level: number, index: bigint) => `${name}:${level}:${index}`;
 const encodeMeta = (root: Buffer, depth: number, size: bigint) => {
   const data = Buffer.alloc(36);
   root.copy(data);
@@ -34,13 +34,13 @@ export abstract class TreeBase implements MerkleTree {
   protected cachedSize?: bigint;
   private root!: Buffer;
   private zeroHashes: Buffer[] = [];
-  protected cache: { [key: string]: Buffer } = {};
+  private cache: { [key: string]: Buffer } = {};
 
   public constructor(
     protected db: LevelUp,
     protected hasher: Hasher,
-    protected name: string,
-    protected depth: number,
+    private name: string,
+    private depth: number,
     protected size: bigint = 0n,
     root?: Buffer,
   ) {
@@ -188,7 +188,7 @@ export abstract class TreeBase implements MerkleTree {
    * @returns The latest value at the given index.
    * Note: If the value is not in the cache, it will be fetched from the database.
    */
-  protected async getLatestValueAtIndex(level: number, index: bigint, includeUncommitted: boolean): Promise<Buffer> {
+  private async getLatestValueAtIndex(level: number, index: bigint, includeUncommitted: boolean): Promise<Buffer> {
     const key = indexToKeyHash(this.name, level, index);
     if (includeUncommitted && this.cache[key] !== undefined) {
       return this.cache[key];
@@ -238,5 +238,52 @@ export abstract class TreeBase implements MerkleTree {
     } else {
       await this.db.put(this.name, data);
     }
+  }
+
+  /**
+   * Appends the given leaves to the tree.
+   * @param leaves - The leaves to append.
+   * @returns Empty promise.
+   *
+   * @remarks The batch insertion algorithm works as follows:
+   *          1. Insert all the leaves,
+   *          2. start iterating over levels from the bottom up,
+   *          3. on each level iterate over all the affected nodes (i.e. nodes whose preimages have changed),
+   *          4. fetch the preimage, hash it and insert the updated value.
+   * @remarks This algorithm is optimal when it comes to the number of hashing operations. It might not be optimal when
+   *          it comes to the number of database reads, but that should be irrelevant given that most of the time
+   *          `getLatestValueAtIndex` will return a value from cache (because at least one of the 2 children was
+   *          touched in previous iteration).
+   */
+  protected async appendLeaves(leaves: Buffer[]): Promise<void> {
+    const numLeaves = this.getNumLeaves(true);
+    if (numLeaves + BigInt(leaves.length) - 1n > this.maxIndex) {
+      throw Error(`Can't append beyond max index. Max index: ${this.maxIndex}`);
+    }
+
+    // 1. Insert all the leaves
+    let firstIndex = numLeaves;
+    let level = this.depth;
+    for (let i = 0; i < leaves.length; i++) {
+      const cacheKey = indexToKeyHash(this.name, level, firstIndex + BigInt(i));
+      this.cache[cacheKey] = leaves[i];
+    }
+
+    let lastIndex = firstIndex + BigInt(leaves.length);
+    // 2. Iterate over all the levels from the bottom up
+    while (level > 0) {
+      firstIndex >>= 1n;
+      lastIndex >>= 1n;
+      // 3.Iterate over all the affected nodes at this level and update them
+      for (let index = firstIndex; index <= lastIndex; index++) {
+        const lhs = await this.getLatestValueAtIndex(level, index * 2n, true);
+        const rhs = await this.getLatestValueAtIndex(level, index * 2n + 1n, true);
+        const cacheKey = indexToKeyHash(this.name, level - 1, index);
+        this.cache[cacheKey] = this.hasher.compress(lhs, rhs);
+      }
+
+      level -= 1;
+    }
+    this.cachedSize = numLeaves + BigInt(leaves.length);
   }
 }
